support dry-hopping in the keg
The volume calculations change slightly over fermentor dry-hopping;
we want less beer if dryhopping in the keg, not more like when
hopping in the fermentor.

Currently, the software will not take into account the volume that the
keg dryhops use, but it will warn if it thinks they are >1dl in volume.
Since we cannot know the volume of the hop container, and the relative
volume is rather small (<1% in all sensible cases), we'll leave it as
the responsibility of the brewer to decide what will fit into the keg
and what won't.