support adding sugars into boil/fermentor
NOTE: the calculations are not strictly correct; adding the same
amount of sugar into the ferment vs. into the boil increases
the strength more due to smaller volume (kettle loss has been
dealt with).  that said, it's a small error, and try to solve
for the difference really made my head hurt.  so, maybe later,
but good enuf for now.