rework how recipe volumes function
Recipes which specify all inputs in to-be-resolved units (e.g. percent
grains + target strength, hops-by-IBU) *may* specify an inherent volume.
Recipes which specify any of those inputs by mass *must* specify an
inherent volume.

If -v is specified on the wbctool command line, it overrides the inherent
volume of the recipe.  For recipes with inputs as absolute masses,
the masses are scaled proportionally to the inherent volume.

For example, if a recipe's inherent volume is 10l and it includes 5kg
of a certain grain, wbctool run without -v will output a sheet which
uses 5kg of grain, but -v 20l will use 10kg of grain.

NOTE: scaling a per-absolute-mass recipe will *NOT* give you x times the
exact same beer due to variances in wort strength, proportional boiloff
rate and losses.  The differences can mostly be ignored for homebrewing
scales; just something to be aware of in case your homebrewing batches
range from 1gal to 1bbl...