fermentables: print 100% efficiency extract and predicted extract
The new information of 100% efficiency extract (i.e. derived from the
figure reported by maltster) gives the brewer an easy way to calculate
conversion efficiency and lauter efficiency, as opposed to just staring
at the mash efficiency as a whole.

The old field of predicted final strength per extract component was
useless except for non-amusing amusement, and was retired.