Add support for specifying mashin ratio as total water percentage.
For example, setting it to 50% will give the best batch sparge
efficiency (assuming a single sparge, of course).

I'm not really sure this value needs to be a recipe paramater and not a
system parameter.  Some further meditation is required to decide where
the parameter will ultimately live, but let's dump it here for now.