Add support for "steep" fermentables.
It's currently intended for dark grains which the brewer can add
sometime after mashin and remove before the boil.  For example, the
"steep" grains can be added to the top of the mash during lautering,
or in a grain bag into the already collected wort.  For the purposes
of the software, the steep fermentables do not affect strike water
temperature or volume, but do affect preboil volume and strength.

Technically, the "steep" phase is like steeping grains in extract brewing,
though the software does not (yet?) support a brew without a mash stage.