Add an "experimental" -m[iniprint] argument to wbctool
If miniprint is specified, some information will be left out of the
print, such as ABV speculation.  The idea is that during recipe
formulation you'd not specify -m to get a better idea of the result,
and when you print the thing out for brewing, you'd leave out the
now-"useless" info.  (for example, since you don't know the initial
gravity for sure until after you brew, it's much better to use
wbcabv for that than rely on printed calculations with a potentially
invalid premise)

essentially, conserves a bit of paper space.  fitting on one side
of one piece of paper is still the holy grail