Split result printing from Recipe into a separate module
The point is the breathe some more structure into the program, and also
to make sure we keep honest and actually do all calculations before the
output phase.

Included in the commit is a bunch of smaller changes that weren't
worthwhile to split into separate commits from a messy tree at
this pre-alpha stage of development.