add wbcextraext for calculations on extract into existing wort
The tool is useful for calculations based on values that are not
entirely known at recipe formulation time (exact strength, volume,
attenuation, ...), and can be used for more precise insight on
final strength, abv, predicted gravity after the addition is
fully fermented, etc.

The tool takes original strength, volume, and amount of extract
to be added.  The amount of extract can be specified as:
  * mass (useful when adding pure sugar or LME)
  * mass @ extract % by mass (useful e.g. for honey or LME)
  * volume @ strength (useful when adding a wort or starter)

Optionally, using the -f parameter you can supply the final strength
of the original wort, and the tool will calculate the aggregate final
strength.  You can also supply the expected real attenuation of the
addition.

So, for example, if you have 5gal of 1.060 wort which fermented down
to 1.015, and you add a pound of honey (75% extract, of which 95% is
fermentable) to it, you can model that by:

	wbcextraext.py -f 1.015,95% 1.060 5gal 1lb@75%

And get the following result:

=====
Strength, Original          :        14.7°P         1.060
Strength, Aggregate         :        16.1°P         1.066

Extract, Original           :       2.96 kg      6 1/2 lb
Extract, Added              :         340 g      12.00 oz
Extract, Aggregate          :       3.30 kg      7 1/4 lb

Water, Original             :      17.11 kg   37 11/16 lb
Water, New                  :      17.22 kg   37 15/16 lb

Volume, Original            :         18.9l        5.0gal
Volume, Aggregate           :         19.3l        5.1gal

Final Strength (95% add RA) :         3.4°P         1.013
=====

And then, if you like, you can use wbcabv to figure out the
honey-included alcohol content (which is 5.9% without the
honey and 7.0% with the honey, so not a trivial difference).