Add support for opaques.
They are additions which do not affect calculations (e.g. spices).

The main idea is the ability to add "notes" to especially the boil,
so that one doesn't have to keep track of timers outside of the
brewday dump.

There are two types of opaques: ones specified by mass, and ones
specified by "opaque".  Mass can be given either in absolute mass, or
mass/batchvolume.  Opaques by mass will be scaled.  The other option is
fully opaque, which can be any specifier, e.g. "1 tablet" for whirlfloc.
This quantity will not be interpreted or scaled, and will be printed
exactly as is.

TODO: boil additions of fermentables are not printed in the "boil"
section, so one still must keep separate timer counters -- boil
fermentables are not even given a time in the recipe input. Since
what used to be the "Hops" printout now includes items beyond hops,
should *all* additions (including all fermentables) be dumped there?
Double-dumping of all fermentables would cause the "standard" printout
to overflow one page, so needs more consideration.