Get rid of diastatic power spec for fermentables.
Some spec sheets don't include values at all and some are in the
"hard-to-believe" range.  Furthermore, the only thing the software
could do with the value is estimate if the mash will convert, but I've
never been in a situation where I wondered if the mash would convert.
So, getting rid of the field simplifies inputting new fermentables.