Make fermentables take the same timespecs as everything else
Essentially, this will allow us to print the fermentables as part
of "additions", which will help not forgetting that 10min boil
addition of sugar.  This change is also part of a larger cleanup
to make the Recipe internally use Worter for everything.

There are still two sets of stages, but the separation between them
is clearer.  Timespec.stages includes possible addition stages,
such as kettle or mash.  Worter.stages is for stages when we want
to measure the strength/volume of what we're doing, e.g. preboil
or fermentor.  (Technically the stages should be on top of those
classes, not in them, but keep them in there for now -- you don't
have to use them unless you want to)

The input-side change is that all fermentables in recipes must now
be changed to accept new timespecs, e.g. "boil" must be changed to
"10min" (or whatever), "steep" must be changed to "mash @ steep",
and so forth.

The only output change is that there is an extra separator printed
between mash and steep/sparge additions (e.g. phosphoric acid).
That's because the addition stages are now defined by Timespecs, and
it's cleaner to separate "mash" and "parallel to mash" additions into
separate timespecs.  We need to do the separation because otherwise the
mash water calculations will be wrong.

Also, add an "undef" range for the Fermentor timespec, so that you
don't have to specify when you want to add fermentables (you can, if
you want to).