wbcextraext: accept multiple adjustments
The status after each adjustment will be printed.

The aggregate results are the aggregates for all previous additions with
the originals being the aggregate values from the previous stage.
So, for example, if you add 5kg sugar, 2l water and 1kg sugar, you
can now just give "5kg 2l@1.000 1kg" as the parameters instead of
having to manually copy the results and invoke the program again.