Use Worter internally.
Simplifies and corrects a lot of calculations.

This is a "mega-commit" which touches pretty much everything in WBC.
and recipe.  There are changes to every test-recipe output, but they
are minor changes (some percent), and [hopefully] the new results are
more accurate.

The main user-visible change is that grain absorption is now the
*real* absorption (*not* apprent absorption, i.e. water_in - wort_out).
The default config value for MLTs has been updated from 1.05L/kg to
1.50L/kg to reflect the change in calculation.

The stolen wort feature is currently broken, but hopefully will make
a comeback soon.