add the wbcsolve utility
Takes two of {volume,extract,strength} and calculates the third.

The utility can be used to calculate things such as "what strength will
I get from 5kg of 80% honey for 20l" (wbcsolve 5kg@80% 20l) or
"what's the ppg of dextrose which is 91% sugar" (wbcsolve 1lb@91% 1gal)

Note: this tool deals with the final volume of the solution.  If you
are intested in what happens when you add a certain volume of water,
use wbcadjust.