Allow recipes without a mash.
There's still a bit of wonkiness in the output for non-beer,
e.g. pre/postboil is always printed, along with IBUs and total
hop use.  Regardless, recipes that include only fermentor/package
fermentables now work.

Also, you can specify weird things, such as acid additions
into the mash in a recipe without a mash.  I'm not convinced that
it's worth it to throw an error for those, i.e. garbageingarbageout.