recipe: calculate the "extract-equivalent" going into packaging
Here, "extract-equivalent" means both unfermentable extract and
the extract that was fermented.  In other words, it's the total
extract minus mash, kettle and fermentor losses.

Previously the last column under Fermentables used to be the
amount of extract multiplied by the mash efficiency for grist,
and the straight-up the amount of extract for everything else.
Now you can track e.g. how much extract you lose if you add
1kg of sugar into the boil vs. into the fermentor.