Stop carrying the user-supplied search-name for fermentables
Unifies the "signature" of fermentable additions with
everything else: (object, amount, timespec).

The name was carried for the purpose of printing the name of the
fermentable with the same capitalization as with which it was typed in
the recipe.  If you for some reason still want an alternatively capsed
printout for a fermentable, define an alias.