hops: add T45/T90/plug types (and stop accepting pellet)
Also, push the name formatting from output into Hop.

Upcoming, we'll get rid of hop-specific AA% field in output,
and integrate that into a nicely formatted name field.

Moving from "pellet" to more specific types has been on the radar
for a while, and now it was prompted by the trivial matter of
"pellet" being 6 chars while the rest are 3/4 (we are a bit
horizontally challenged in printable 80col output)

Plug/T45/T90 are currently treated in the code the same as pellet
was (i.e. different from "leaf").  In the future, "wet" will also
be added, but since that requires different code treatment, put it
on the TODO for now.