additions & hops: format AA% into name
Been playing with the new output format for some whiles.  I'm still
not sure I like it, but maybe it just looks different from the old
for now and I'll get used to it.  At least the type of hop is now
nicely columnized, and the space occupied by the AA% field is available
to other types of additions as a longer namelen.