recipe: we need 3(!) volumes: inherent, set, scaled
If we conflate inherent and set (non-scaled), we can't sanity-check
for a recipe spec with absolute units but without an inherent volume.

I somehow thought I'd already solved this, but apparently just
punted.  I'm not sure we need a separate scaled and non-scaled
version of setting the volume.

Further updates on this trivial but complexified matter promise to
be legion (= job security 101 for software engineers).