Introduce additions.
An addition is anything that gets added to a recipe.  It involves
an amount, a time, and the object of what gets added (e.g. hop).
An optional resolver for the amount can be given, so that e.g. scaling
and m/v additions "just work" without every caller fudging with
them.  Incidentally, the change fixes the scaling of bymass and
byvolume opaques.

Hops and opaques were converted to use additions, with fermentables
hopefully following soon.  Incidentally, the goal of the whole operation
is to make fermentables printable also in the "additions" section,
especially the ones that get added to the boil.