Print the type of each addition (output_text).
The type is printed as the first char of the line.  Teaches
output_text about various addition types, but the alternative
is the addition type knowing what is happening in the output
routine, so this is the smaller of two evils.  Tried to pick
visually distictive characters, so not e.g. (h)op, (f)ermentable.