Add a very preliminary addition type "water"
Currently it functions like an opaque, but with the exception
that the m/v and v/v calculations are performed on the volume
at the stage instead of the final volume.  E.g., a 0.1g/L addition
to mashin will result in 0.1g being multiplied by the strike water
volume, not the final volume.

Also, uses the old "water" keyword, so water notes are now dead.
Use "notes" instead for any content you want carried over.

to be refined ...