wbcsolve: add optional starting sauce parameter
The starting sauce is specific as -s 'volume @ strength'.

The starting sauce can be used to solve things like
"how much water do I need to add to dilute this to 10Bx?"
or "how much 83% honey do I need to add to boost it to 1.090?"

Also introduce a "minvol" value for volume, which solves for the minimum
volume, and is useful in cases where you want to strenghten or dilute and
not hit any specific volume. (minvol works only with starting sauce).