Introduce the optional output_text-pagelen sysparam
Can be used to control the output page length.

If specified, will output a pagefeed if the next output section,
where section is e.g. Fermentables, Additions, ..., would
exceed "pagelen" lines.  In essence, when compiling a recipe
that's going to be printed on paper, the parameter can be specified
to make page changes occur at output boundaries.

XXX: the parameter is specific to output_text, yet sysparams knows
about it.  sysparams probably needs a touch-up so that it doesn't have to
know about optional module-specific output parameters; consider e.g. the
case where a 3rd party output_html module appears in the future