user-visible changes to WBC fermentor timespec
The timespec was originally for dryhopping, make it more general.

Reverse the order.  Previously, "day 0" was considered the transfer day
to packaging and "day n" was the first day of an n-day fermentation.
For long-fermenting things such as meads and wines it's inconvenient
to write recipes with e.g. early nutrient additions has happening on
day 300.  So, now daycount means the number of days that fermentation
has been ongoing.

Also, allow unspecified "outdays" even if "indays" is specified.
Unlike dry hops, things such as nutrients don't come out.  And even
with dry hops is very rare to pull them out except as part of the
transfer...