Account for fermentable moisture in calculations.
This has a few-dL impact on the water needed for all-grain mashes,
but has an up to several liter impact on for example meads (honey being
~20% water).

Test output changes, and is a few-dL reduction of water, plus roundoff
errors (0.1dL water in mash steps, 0.1degP, 0.0x kg fermentable/extract,
etc).

TODO: add moisture to some fermentables (previously it was used to
calculate only extract for DB).  also, fix the mash to not think
that the moisture added as part of the grains is being added
as water.  So, essentially this commit checkpoints test output.