Do not count moisture in the grains as water added to the mash
Now the total calculated water in the recipe matches the sum of
the mash and sparge waters (well, assuming there was no boilvol
limitation ... and assuming the recipe includes a mash).

The mash code still ignores the moisture in the grains for e.g.
the mash thickness and the thermal properties.  It's on average
"correct enough" and doesn't currently warrant fixing and the
added complexity from such a fix.