add date sysparam, wbc_datefmt utility, minor output_text changes
In essense, these changes make wbcrecipe output include the
date and allow the user to override it with an arbitrary string.
In case the wish is to be consistent, the user can format
that arbitrary date in a consistent fashion with the
format string used by WBC when no user-date is specified.

The rationale is that 1) saves lines over the previous
"best practice" to use the date produced by enscript
2) in case the printout was produced on a different day from
brewday, it is possible to reflect the true brewday in the
printout.

example: wbcrecipe -P dt="$(date +"$(wbc_datefmt)" -d '2 days ago')" ...
(or just type the date on the command line manually if nested
execs aren't a thing of beauty for you)