Add initial support for liquid fermentables.
Liquid fermentables may be specified by volume, volume/volume,
or percentage.  Percentage is used as a mass of the total,
like with other fermentables.  Liquid fermentables are still
displayed as volume in the output.

Add a "maximum" strength specifier, which means that the
strength will be whatever the fermentables and final volume
result in with no added water.  This specifier can only
be used with a significant amount of liquid fermentables, i.e.
the sauce has to be "a liquid".

Liquid fermentables do not yet work in all cases, such as for
a graf, but something like an apple cider with priming sugar
or apple wine is possible.

Add "apple juice" (12 Brix) as the canonical
liquid fermentable example.

This support is still going to be revamped with the long-looming
fermentables & ingredients rototill.  For one, I doubt that
everyone's apple juice is 12 Brix ...