Optionally output-text the color estimate as an actual color
Controlled with the "output_text-color" ("oC") sysparam, default off.

Works only on terminals which support 24bit ANSI colors.  It is
up to the user to "know" if the underlying terminal supports them;
test once, and decide if the result is something you want to see.

The feature works only for recipes where the color is estimated (beers).

Meant to give a rough idea, and not be a perfect representation of what
the beer looks like in the glass you're drinking it from in the given
ambient lighting.  Notably, the background color of the terminal also
affects the perception of the color: black colors don't look fully
black against a black background, but they do so against a white
background.