wbcrecipe: add -i[ncludeyaml] and -I[nlineyaml] options
Makes it possible to augment the recipe without having to
edit a monolith.

For example, if you add the following to your wrapper script:

	-I "opaques: { byopaque: [ [ chiller, '', 10min ] ] }"

A timer for a 10min "addition" of a chiller will be automatically
included in all recipes.  (potential caveat: works only with
recipes with a >=10min boil)

Note: if you want to add once-off "leftover" fermentables, e.g.
-I 'fermentables: { mashin: { somemalt: 100g } }' to some
recipe which does not specify an inherent volume, you need to specify
the volume with -I 'volume: x' instead of -v.

This feature might still undergo some changes (includepaths,
carrying over "defs", etc.  not sure yet).