Mash: allow multiple heating styles in the same mash.
In other words, can now infuse, heat and decoct in the same recipe.

The main user-visible change is that mlt_heat=direct no longer
implies direct-fired mashes, the default is infusion.  (I will
rename the mlt_heat sysparam soon, probably to mashtun_striketemp)

I want to change the mashparams syntax eventually, but for now
the magic separator is ";", i.e. a step is defined as:

	time @ temperature ; method

time and method are optional, with time defaulting to undefined,
and method to infusion as mentioned above

Replace the infusion,step test logo with a mish[mash],step lego.
The latter includes all of strike, decoction, infusion and direct fire.