WBC: tighten volumetric iteration precision from 1dL to 1cL
The change is mostly cosmetic, but avoids the displayed final volume
being 0.1L off of what the user specified in cases where the roundoff
is unfortunate.  Results in a bunch of other roundoff-related cosmetic
changes, such as a 0.1g difference in hops or a 1g difference in extract.

Maybe should make the required precision a percentage instead of an
absolute number, but until someone wants to use WBC for non-homebrew
sized batches, not going to bother.