Change water additions to use the volume in most cases.
Previously the amount of water at the stage was used.

The rationale is that if you specify an addition of 0.1g/L
and your volume happens to be 20L, it looks plain odd to
see, e.g., 1.8g calculated.

There is now an inconsistency with the mash, as in the addition there
is for the water instead of the hypothesized converted wort volume, but
I'm guessing that this model is the most intuitive and expected by users.

In essence, now the calculations are:
  1) water volume for timespecs where water is added (e.g. mash, sparge)
  2) total volume for other timespecs (e.g. boil, fermentor)