wbcrecipe: allow to input fermentation notes for a recipe
Arguably, they're a very key part of a recipe's outcome.

It's an opaque string that is meant to contain the estimate for
temperatures, days, etc. of fermentation.  Technically, a fermentation
note is no different from a recipe note, but having a separate field
for it allows us to output either the given schedule, or if none given
in the recipe, an empty field that can be scribbled in by hand later.

The method of input is to make "yeast" accept a two-element array,
where the first element is the yeast string as previously, and the
second element is the fermentation schedule string.  If yeast is given
as a string instead of an array, the fermentation schedule is taken
as blank.  The reason for overloading "yeast" was twofold.  First, the
fermentation schedule is always tied to the yeast, so it makes sense to
group them together.  Second, and maybe slightly silly, using "yeast"
as the keyword allows all ``one-liner'' toplevel info fields to remain
<= 6 chars, and allow for a sofftab of 8 -- "fermentation", "ferment",
"zymosis", etc. would all be a minimum of 7 plus colon = >=8.

Notably, anything that isn't a recipe detail is better served by brewday
notes.  For example, using a yeast cake for a high-octane beer is a
recipe detail, but where the yeast cake came from is a brewday note.

Example:

yeast: [ 1yeast, pitch at 20; free rise to 25 by day 3 ]

For completeness, the following still works: "yeast: 1yeast", and
the output_text fermentation plan field is empty.

Also, shuffle output_text so that "Yeast" and "Fermentation plan"
are close to each other.  (add the word "plan" to make the
field more visually distinctive from "Recipe notes", which is the
same length as "Fermentation").