WBC: add initial support for nutrients and more specifically YAN
The amount of YAN in a recipe is tracked and nutrient additions are
calculated according to the specified YAN level.  A nutrient is specified
as [ name, YAN-by-mass% ] (liquid nutrients are currently not supported).
A nutrient addition is specified as [ nutrient, amount, timespec ].

The amount can be specified as m, m/v, m/(v*degBx), and any of the
previous as a recipe total.  If the flag "gross" is specified in the
amount field, the mass is for the gross mass of the nutrient, otherwise
YAN; this flag is useful for complex nutrients (= not DAP) which you may
wish to specify using the gross amount instead of the YAN contribution.
For example:

yan:
  - [ [ _est from must_, 100% ],  50mg/L,                day 0 ]
  - [ [ Fermaid-K,        10% ],  5g/10L gross,          day 2 ]
  - [ [ DAP,              21% ],  Recipe 7.5mg/L perBx,  day 4 ]

The above specifies that there is an estimated 50mg/L YAN from the
must, an addition of 5g/10L Fermaid-K on day 2, and a total of 7.5mg/L
per degree Brix which is to be satisfied using DAP.  Notably, the way
to specify the amount of YAN from the must is with a pseudo-addition.
Equally notably, the 7.5mg/L perBx corresponds to a low-nitrogen yeast,
and you would use 9mg/L perBx for a yeast with medium nitrogen needs etc.

The volume used in the calculation is the volume in the fermentor,
not the final volume like with most other additions.

Given that IBU and YAN are unlikely to be used in the same recipe,
this feature hijacks the IBU output space in the recipe header
and the additions column.